requirements-gpu.txt
fastapi
uvicorn
pillow
rembg
onnxruntime-gpu
python-multipart